flatten slr tests
simplified all the tests and the contexts
