+++ src/NServiceBus.AcceptanceTests/Recoverability/Retries/When_performing_slr_with_serialization_exception.cs
@@ -1,15 +1,28 @@
 ﻿namespace NServiceBus.AcceptanceTests.Recoverability.Retries
 {
+    using System;
+    using System.Linq;
     using System.Threading.Tasks;
     using NServiceBus.AcceptanceTesting;
+    using NServiceBus.AcceptanceTests.EndpointTemplates;
+    using NServiceBus.Config;
+    using NServiceBus.Features;
+    using NServiceBus.MessageMutator;
     using NUnit.Framework;
 
-    public class When_performing_slr_with_serialization_exception : When_performing_slr
+    public class When_performing_slr_with_serialization_exception : NServiceBusAcceptanceTest
     {
+        public class Context : ScenarioContext
+        {
+            public byte OriginalBodyChecksum { get; set; }
+            public byte SlrChecksum { get; set; }
+            public bool ForwardedToErrorQueue { get; set; }
+        }
+
         [Test]
         public async Task Should_preserve_the_original_body_for_serialization_exceptions()
         {
-            var context = await Scenario.Define<Context>(c => { c.SimulateSerializationException = true; })
+            var context = await Scenario.Define<Context>()
                 .WithEndpoint<RetryEndpoint>(b => b
                     .When(bus => bus.SendLocal(new MessageToBeRetried()))
                     .DoNotFailOnErrorMessages())
@@ -18,5 +31,93 @@
 
             Assert.AreEqual(context.OriginalBodyChecksum, context.SlrChecksum, "The body of the message sent to slr should be the same as the original message coming off the queue");
         }
+        public class RetryEndpoint : EndpointConfigurationBuilder
+        {
+            public RetryEndpoint()
+            {
+                EndpointSetup<DefaultServer>(configure =>
+                {
+                    configure.DisableFeature<FirstLevelRetries>();
+                    configure.EnableFeature<SecondLevelRetries>();
+                    configure.EnableFeature<TimeoutManager>();
+                    configure.RegisterComponents(c => c.ConfigureComponent<BodyMutator>(DependencyLifecycle.InstancePerCall));
+                })
+                .WithConfig<SecondLevelRetriesConfig>(c => c.TimeIncrease = TimeSpan.FromMilliseconds(1));
+            }
+
+            static byte Checksum(byte[] data)
+            {
+                var longSum = data.Sum(x => (long)x);
+                return unchecked((byte)longSum);
+            }
+
+            class BodyMutator : IMutateOutgoingTransportMessages, IMutateIncomingTransportMessages
+            {
+                Context testContext;
+
+                public BodyMutator(Context testContext)
+                {
+                    this.testContext = testContext;
+                }
+
+                public Task MutateIncoming(MutateIncomingTransportMessageContext transportMessage)
+                {
+                    var originalBody = transportMessage.Body;
+                    testContext.OriginalBodyChecksum = Checksum(originalBody);
+                    var newBody = new byte[originalBody.Length];
+                    Buffer.BlockCopy(originalBody, 0, newBody, 0, originalBody.Length);
+                    //corrupt
+                    newBody[1]++;
+                    transportMessage.Body = newBody;
+                    return Task.FromResult(0);
+                }
+
+                public Task MutateOutgoing(MutateOutgoingTransportMessageContext context)
+                {
+                    return Task.FromResult(0);
+                }
+            }
+
+            class ErrorNotificationSpy : IWantToRunWhenBusStartsAndStops
+            {
+                Context testContext;
+                BusNotifications notifications;
+
+                public ErrorNotificationSpy(Context testContext, BusNotifications notifications)
+                {
+                    this.testContext = testContext;
+                    this.notifications = notifications;
+                }
+
+                public Task Start(IBusSession session)
+                {
+                    notifications.Errors.MessageSentToErrorQueue += (sender, message) =>
+                    {
+                        testContext.ForwardedToErrorQueue = true;
+                        testContext.SlrChecksum = Checksum(message.Body);
+                    };
+                    return Task.FromResult(0);
+                }
+
+                public Task Stop(IBusSession session)
+                {
+                    return Task.FromResult(0);
+                }
+            }
+
+            class MessageToBeRetriedHandler : IHandleMessages<MessageToBeRetried>
+            {
+                public Task Handle(MessageToBeRetried message, IMessageHandlerContext context)
+                {
+                    return Task.FromResult(0);
+                }
+            }
+        }
+
+        [Serializable]
+        public class MessageToBeRetried : IMessage
+        {
+        }
+
     }
 }
